@@ -35,7 +35,7 @@
   },
   "lint": "eslint src/*",
   "devDependencies": {
-    "eslint": "^5.3.0",
+    "eslint": "^5.6.1",
     "eslint-config-airbnb": "^17.1.0",
     "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-jsx-a11y": "^6.1.2",
