@@ -1,20 +1,20 @@
-import React, { Component } from "react";
-import { ChatBox, ChatInput } from "./index";
-import { firestore } from "../../fire";
-import Login from "../authentication/login";
-import BottomNav from "../BottomNavigation/BottomNav";
-import { thunkLogInUser, thunkLogOutUser, actionSetTipDestination} from "../../state/user/reducer";
-import { connect } from "react-redux";
-import BlockChainBar from "../ethereum/BlockChainBar";
-var firebase = require("firebase");
+import React, { Component } from 'react';
+import { connect } from 'react-redux';
+import { ChatBox, ChatInput } from './index';
+import { firestore } from '../../fire';
+import Login from '../authentication/login';
+import BottomNav from '../BottomNavigation/BottomNav';
+import { thunkLogInUser, thunkLogOutUser, actionSetTipDestination } from '../../state/user/reducer';
+import BlockChainBar from '../ethereum/BlockChainBar';
 
+const firebase = require('firebase');
 
 class ChatBucket extends Component {
   constructor(props) {
-    
     super(props);
     this.state = {
-      messages: []
+      messages: [],
+      chatOpen: false
     };
 
     this.getInitialMessages = this.getInitialMessages.bind(this);
@@ -24,18 +24,17 @@ class ChatBucket extends Component {
 
   async componentDidMount() {
     const { discourseId } = this.props;
-    
+
     this.subscribeToMessageUpdates(discourseId);
   }
 
   async getInitialMessages(discourseId, limit = 50) {
-    
     const messages = await firestore
-      .collection("discourseList")
+      .collection('discourseList')
       .doc(discourseId)
-      .collection("messages")
+      .collection('messages')
       .limit(limit)
-      .orderBy("timestamp")
+      .orderBy('timestamp')
       .get();
 
     messages.forEach((message) => {
@@ -47,22 +46,21 @@ class ChatBucket extends Component {
     const { messages } = this.state;
     // let newMessageState = new Set([...messages, message]);
     // let messageState = Array.from(newMessageState);
-    const newMessages = [ ...messages, message ]
+    const newMessages = [...messages, message];
     this.setState({ messages: newMessages });
   }
 
   async subscribeToMessageUpdates(discourseId) {
-    
     const { messages } = this.state;
     await firestore
-      .collection("discourseList")
+      .collection('discourseList')
       .doc(discourseId)
-      .collection("messages")
-      .orderBy("timestamp", "desc")
+      .collection('messages')
+      .orderBy('timestamp', 'desc')
       .limit(100)
       .onSnapshot((snapshot) => {
         snapshot.docChanges().forEach((change) => {
-          let messageArray = messages;
+          const messageArray = messages;
           messageArray.push(change.doc.data());
           this.setState({ messages: messageArray });
         });
@@ -72,64 +70,72 @@ class ChatBucket extends Component {
   // This method for TimeStamp is INSECURE
   postMessage(message) {
     const { user, discourseId } = this.props;
-    const { displayName, uid, photoURL } = user
+    const { displayName, uid, photoURL } = user;
     const date = new Date();
     const messageObj = {
       body: message,
       userName: displayName,
       timestamp: date,
-      uid: uid,
-      photoURL: photoURL,
-    }
+      uid,
+      photoURL,
+    };
 
-    this.addSingleMessageToState(messageObj)
+    this.addSingleMessageToState(messageObj);
     firestore
-      .collection("discourseList")
+      .collection('discourseList')
       .doc(discourseId)
-      .collection("messages")
+      .collection('messages')
       .add(messageObj)
-      .then(function(docRef) {
-        
-      })
-      .catch(function(error) {
-        console.error("Error adding document: ", error);
+      .then((docRef) => {})
+      .catch((error) => {
+        console.error('Error adding document: ', error);
       });
   }
 
   render() {
-    
-    
     const { messages } = this.state;
-    const { logOutUser, logInUser, isLoggedIn, displayName, discourseId, tipDestination } = this.props;
+    const {
+      logOutUser,
+      logInUser,
+      isLoggedIn,
+      displayName,
+      discourseId,
+      tipDestination,
+    } = this.props;
     return (
       <div className="Chatbucket-container">
-      <BlockChainBar />
-        {isLoggedIn ?
+        <BlockChainBar />
+        {isLoggedIn ? (
           <div onClick={() => logOutUser()}>Logout</div>
-        :
+        ) : (
           <div onClick={() => logInUser()}>Login</div>
-        }
-        <ChatBox msgArray={messages} setTipDestination={this.props.setTipDestination}/>
+        )}
+        <ChatBox
+          msgArray={messages}
+          setTipDestination={this.props.setTipDestination}
+        />
 
         {isLoggedIn ? (
-          <ChatInput postMessage={this.postMessage} discourseId={discourseId} />
+          <ChatInput
+            postMessage={this.postMessage}
+            discourseId={discourseId}
+          />
         ) : (
           <div onClick={() => logInUser()}>Login</div>
         )}
         <div className="bottom-nav">
-          <BottomNav />
+          
         </div>
-
       </div>
     );
   }
 }
 
-//CONTAINER====================================================================
+// CONTAINER====================================================================
 function mapState(state) {
   return {
     user: state.userReducer.user,
-    isLoggedIn: state.userReducer.isLoggedIn
+    isLoggedIn: state.userReducer.isLoggedIn,
   };
 }
 
@@ -142,14 +148,14 @@ function mapDispatch(dispatch) {
       dispatch(thunkLogInUser());
     },
     setTipDestination: (destination) => {
-			dispatch(actionSetTipDestination(destination));
-		}
+      dispatch(actionSetTipDestination(destination));
+    },
   };
 }
 
 ChatBucket = connect(
   mapState,
-  mapDispatch
+  mapDispatch,
 )(ChatBucket);
 
 export default ChatBucket;
